Added loci view, which shows the paths each point moves along
It was necessary to add a non-rounding method to glyphApplyVariations() in samsa-core.js. Without this, drawing the path by means of multiple line segments leads to aliasing because the resolution of the point locations is in font units, so limited by unitsPerEm (typically 1000 or 2048). This non-rounding method is only used for calculating the loci.