Added an MP4 exporter, but inactive button until we can work out Safari, Chrome bugs
Main addition is a export an MP4 movie that traverses the font in a certain way. Creating an MP4 movie in-browser is hard, so we export lots of PNG frames (samsa-frame000.png…samsa-frame<n>.png) as well as a shell script (samsa-movie.sh) that invokes ffmpeg to compile the frames into a movie. Export of each frame is invoked by triggering a click event on a temporary <a> element, with "href" as the data url of the PNG file and "download" as the frame’s filename. Initial traversal is simply to animate axis 0 from min to max in 50 steps.

The method works well in Firefox, but Safari and Chrome drop most frames. Causes? Maybe a kind of throttling of multiple click events or multiple downloads. Maybe an async bug.