avar: now visualized amd processed properly
samsa-core.js finally handles avar mappings correctly. There was an important bug where a loop was not ending after a match. Now it exits the loop properly.

samsa-gui.js now visualizes avar. Probably needs further work, and maybe a full visualization of the mappings from user -> normalized -> avar’d.

Changes to the "Tuple Variations" panel: renamed it to "Delta sets", and relabelled the "IUP" column to "points", and listed the number of points involved in this delta set. (It previously listed an erroneous number of the number of points moved implicitly by IUP.) Clicking on this number still brings up an alert showing the actual deltas of each point. This will be changed so they are listed in place in the panel.