glyphApplyVariations(): don’t keep instance glyph within the instance object
This was not necessary. Routines that want to keep the glyph around must do so explicitly. This impoves speed everywhere and in particular helps huge fonts. The 65535-glyph 2-axis font JingXiHei-VF_65535.ttf (56 MB) now instantiates as a a 27 MB static font in 4477 ms (2.6 Ghz MacBook Pro Intel Core i7 mid-2012).