Optionally generates uncompressed glyf table (faster)
CONFIG.glyf.compression (true or false) in samsa-core.js controls glyf compression.

In TrueType fonts, each coordinate (x or y) of each point is a 16 bit integer, stored as the delta between itself and the previous point. Compression can occur when the point is the same as the previous point (we store 0 bytes), or if the absolute difference is <256 (we store 1 byte). The compression=false option always uses 2 bytes, so we don’t need to precompute anything and can write flags, x and y all in the same loop.

CONFIG.glyf.compression is recommend to be true if space is the overriding concern. Note that woff2 compression generates identical woff2 files, whether the glyf table uses compression or not.

CONFIG.glyf.compression is recommend to be false if speed is the overriding concern, particularly when buildings fonts in memory, for example in a polyfill setting.

Bahnschrift-ship.ttf (2-axis) produces instances of ~109kb compressed, ~140kb uncompressed.