Switched to a new samsa-config.js model for customizing Samsa's font list etc
The idea is for use it for user-custimizable things like the font list, UI colours, etc., panel order. It uses a new file "samsa-config.js". This also handles data URLs for fonts nicely, base64 encoded. One of the use cases is a Glyphs/Fontlab/Robofont extension to export a custom samsa-config.js file (along with other Samsa files) so that users can immediately test their font in Samsa by double-clicking on samsa-gui.html (running it in the file: protocol): samsa-fonts.json could never be loaded in the file: protocol, but JS files are fine.