Allow "Delta sets" panel content to scroll in x.
In Decovar and other fonts with many axes the "Delta sets" panel
displays a large number of ramps which are not easy to see. Allow the
contents of the panel to scroll. In the future it may be nice to allow
scrolling of the axes only.

The changes to the grid rows prevents the appearance of a vertical
scrollbar when the div containing the svg ramps is larger than 2em,
causing the last delta set representation to overflow the last grid row.