Give the small tables a unique 'data' property, even in arrayBuffer mode
This is efficient because, in arrayBuffer mode, we only create a new DataView, which looks into the existing data.
This allows us to remove the <tableOffset> variable, because it is always 0, which simplifies the code.
Also, the simple tables are given hard-coded offsets, which don't use the <p> variable at all.